refactor(permissions): expose ROLE_ADMIN via lib facade
Callers must import the permissions surface from app.lib.permissions, never
from app.permissions.* internals. Route ROLE_ADMIN through the facade and
switch the user endpoint and create-user CLI to it.
